move rule interface
