+++ src/Contracts/Rule.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Rule;
+namespace ArtARTs36\MergeRequestLinter\Contracts;
 
-use ArtARTs36\MergeRequestLinter\Error\LintError;
+use ArtARTs36\MergeRequestLinter\Linter\LintError;
 use ArtARTs36\MergeRequestLinter\Request\MergeRequest;
 
 interface Rule
