+++ src/Rule/TitleStartsWithRule.php
@@ -2,6 +2,7 @@
 
 namespace ArtARTs36\MergeRequestLinter\Rule;
 
+use ArtARTs36\MergeRequestLinter\Contracts\Rule;
 use ArtARTs36\MergeRequestLinter\Linter\LintError;
 use ArtARTs36\MergeRequestLinter\Request\MergeRequest;
